implemented screen navigation
